Refactor TransactionModal components
