+++ packages/ui/src/common/modals/TransactionModal/MultiTransactionModal.stories.tsx
@@ -2,19 +2,24 @@ import { Meta, Story } from '@storybook/react'
 import React, { ReactNode } from 'react'
 import { interpret } from 'xstate'
 
-import { transactionMachine } from '../model/machines'
+import { ButtonPrimary } from '../../components/buttons'
+import { ModalBody, ModalFooter } from '../../components/Modal'
+import { TextMedium } from '../../components/typography'
+import { transactionMachine } from '../../model/machines'
 
-import { ButtonPrimary } from './buttons'
-import { ModalBody, ModalFooter } from './Modal'
 import { TransactionModal } from './TransactionModal'
-import { TextMedium } from './typography'
 
 export default {
-  title: 'Common/Modals/TransactionModal',
+  title: 'Common/Modals/MultiTransactionModal',
   component: TransactionModal,
 } as Meta
 
-const Template: Story<{ children: ReactNode; state: string }> = ({ children, state }) => {
+const Template: Story<{ children: ReactNode; state: string; steps: { title: string }[]; active: number }> = ({
+  children,
+  state,
+  steps,
+  active,
+}) => {
   const service = interpret(transactionMachine)
   service.start()
 
@@ -30,7 +35,7 @@ const Template: Story<{ children: ReactNode; state: string }> = ({ children, sta
   }
 
   return (
-    <TransactionModal service={service} onClose={() => undefined}>
+    <TransactionModal service={service} onClose={() => undefined} useMultiTransaction={{ steps, active }}>
       {children}
     </TransactionModal>
   )
@@ -49,18 +54,34 @@ const StubBody = () => {
   )
 }
 
+const steps = [
+  {
+    title: 'Bind account for staking',
+  },
+  {
+    title: 'Send a proposal',
+  },
+]
+const active = 1
+
 export const Default = Template.bind({})
 Default.args = {
   children: <StubBody />,
   state: 'start',
+  steps,
+  active,
 }
 
 export const Extension = Template.bind({})
 Extension.args = {
   state: 'extension',
+  steps,
+  active,
 }
 
 export const Pending = Template.bind({})
 Pending.args = {
   state: 'pending',
+  steps,
+  active,
 }
