+++ packages/ui/src/common/modals/TransactionModal/TransactionModal.tsx
@@ -2,47 +2,33 @@ import { useActor } from '@xstate/react'
 import React, { ReactNode } from 'react'
 import { ActorRef } from 'xstate'
 
-import { Modal, ModalHeader } from './Modal'
-import { WaitModal } from './WaitModal'
+import { MultiTransactionModalHeader } from '@/common/modals/TransactionModal/MultiTransactionModalHeader'
 
-export interface TransactionStep {
-  title: string
-}
+import { Modal, ModalHeader } from '../../components/Modal'
+import { WaitModal } from '../../components/WaitModal'
 
-interface MultiTransactionConfig {
-  steps: TransactionStep[]
-  active: number
-}
+import { MultiTransactionConfig } from './types'
 
 export interface TransactionModalProps {
   children: ReactNode
   onClose: () => void
   service: ActorRef<any>
   title?: string
-  asMulti?: MultiTransactionConfig
+  useMultiTransaction?: MultiTransactionConfig
 }
 
-type MultiTransactionModalHeaderParams = {
-  onClick: () => void
-  active: number
-  transactionSteps: TransactionStep[]
-}
-
-const MultiTransactionModalHeader = (props: MultiTransactionModalHeaderParams) => (
-  <ModalHeader
-    onClick={props.onClick}
-    title={`Transaction ${props.active + 1} "${props.transactionSteps[props.active]?.title}"`}
-  />
-)
-
-export const TransactionModal = ({ onClose, children, service, title, asMulti }: TransactionModalProps) => {
+export const TransactionModal = ({ onClose, children, service, title, useMultiTransaction }: TransactionModalProps) => {
   const [state] = useActor(service)
 
   if (state.matches('prepare')) {
     return (
       <Modal modalSize="m" modalHeight="s" onClose={onClose}>
-        {asMulti ? (
-          <MultiTransactionModalHeader onClick={onClose} active={asMulti.active} transactionSteps={asMulti.steps} />
+        {useMultiTransaction ? (
+          <MultiTransactionModalHeader
+            onClick={onClose}
+            active={useMultiTransaction.active}
+            transactionSteps={useMultiTransaction.steps}
+          />
         ) : (
           <ModalHeader onClick={onClose} title={title ?? 'Authorize transaction'} />
         )}
