+++ packages/ui/src/common/modals/TransactionModal/index.ts
@@ -1,0 +1,4 @@
+export * from './TransactionModal'
+export * from './types'
+export { MultiTransactionModalHeader } from '@/common/modals/TransactionModal/MultiTransactionModalHeader'
+export { MultiTransactionModalHeaderParams } from '@/common/modals/TransactionModal/MultiTransactionModalHeader'
