+++ packages/ui/src/common/modals/TransactionModal/types.ts
@@ -1,0 +1,8 @@
+export interface TransactionStep {
+  title: string
+}
+
+export interface MultiTransactionConfig {
+  steps: TransactionStep[]
+  active: number
+}
