+++ packages/ui/src/memberships/modals/UpdateMembershipModal/UpdateMembershipSignModal.tsx
@@ -11,10 +11,10 @@ import { ButtonPrimary } from '@/common/components/buttons'
 import { Label } from '@/common/components/forms'
 import { ModalBody, ModalFooter, Row, TransactionInfoContainer } from '@/common/components/Modal'
 import { TransactionInfo } from '@/common/components/TransactionInfo'
-import { TransactionModal } from '@/common/components/TransactionModal'
 import { TextMedium, TokenValue } from '@/common/components/typography'
 import { useApi } from '@/common/hooks/useApi'
 import { useSignAndSendTransaction } from '@/common/hooks/useSignAndSendTransaction'
+import { TransactionModal } from '@/common/modals/TransactionModal'
 import { metadataToBytes } from '@/common/model/JoystreamNode'
 import { WithNullableValues } from '@/common/types/form'
